fix: refactor duplicated code
