+++ core/src/main/java/co/com/bancolombia/datamask/databind/mask/JsonSerializer.java
@@ -1,6 +1,7 @@
 package co.com.bancolombia.datamask.databind.mask;
 
 import co.com.bancolombia.datamask.cipher.DataCipher;
+import co.com.bancolombia.datamask.databind.util.JsonNodePathUtils;
 import co.com.bancolombia.datamask.databind.util.QueryType;
 import tools.jackson.core.JacksonException;
 import tools.jackson.core.JsonGenerator;
@@ -15,9 +16,6 @@ import java.util.Map;
 
 public class JsonSerializer extends StdSerializer<DataMask<?>> {
 
-    public static final String SEPARATOR = "/";
-    public static final String FORMAT_EXCEPTION = "DataMask only support fields in String format";
-    public static final String NOT_FOUND_IN_TREE = " not found in tree.";
     private final DataCipher dataCipher;
 
     public JsonSerializer(Class<DataMask> dataMaskClass, DataCipher dataCipher) {
@@ -29,19 +27,19 @@ public class JsonSerializer extends StdSerializer<DataMask<?>> {
     @Override
     public void serialize(DataMask value, JsonGenerator generator, SerializationContext provider) throws JacksonException {
         JsonMapper mapper = new JsonMapper();
-        JsonNode objectNode = convertValue(value.getData(), mapper);
+        JsonNode objectNode = JsonNodePathUtils.convertValue(value.getData(), mapper);
         Map<IdentifyField, MaskingFormat> values = value.getFields();
         findFields(objectNode, values);
         generator.writeTree(objectNode);
     }
 
     private void findFields(JsonNode node, Map<IdentifyField, MaskingFormat> maskField) {
-        maskField.keySet().stream().filter((identifyField) ->
-                identifyField.getQueryType().equals(QueryType.PATH)).forEach((identifyField) -> {
-            String[] querySplit = identifyField.getQuery().split(SEPARATOR);
+        maskField.keySet().stream().filter(identifyField ->
+                identifyField.getQueryType().equals(QueryType.PATH)).forEach(identifyField -> {
+            String[] querySplit = identifyField.getQuery().split(JsonNodePathUtils.SEPARATOR);
             this.findFieldForPath(node, identifyField, maskField, querySplit);
         });
-        boolean existFieldsWithQueryTypeName = maskField.keySet().stream().anyMatch((identifyField) ->
+        boolean existFieldsWithQueryTypeName = maskField.keySet().stream().anyMatch(identifyField ->
                 identifyField.getQueryType().equals(QueryType.NAME));
         if (existFieldsWithQueryTypeName) {
             this.findFieldsForName(node, maskField, null, node);
@@ -50,62 +48,16 @@ public class JsonSerializer extends StdSerializer<DataMask<?>> {
 
     private void findFieldForPath(JsonNode node, IdentifyField identifyField,
                                   Map<IdentifyField, MaskingFormat> maskField, String[] querySplit) {
-        var previousContext = node;
-        var nodeParent = node;
-        var pathPart = "";
-        var nodeParentArray = node;
-        var arrayNodeFound = false;
-        var newContext = node;
-        for (String element : querySplit) {
-            pathPart = element;
-            if (!pathPart.isEmpty()) {
-                if (arrayNodeFound) {
-                    nodeParent = findArrayParent(nodeParentArray, pathPart, nodeParent);
-                    previousContext = nodeParent;
-                    newContext = nodeParent;
-                    arrayNodeFound = false;
-                } else {
-                    newContext = findContext(previousContext, pathPart);
-                    previousContext = newContext;
-                }
-                if (!newContext.isValueNode()) {
-                    nodeParent = newContext;
-                }
-                if (newContext.isArray()) {
-                    nodeParentArray = newContext;
-                    arrayNodeFound = true;
-                }
-            }
-        }
+        var pathContext = JsonNodePathUtils.navigatePath(node, querySplit);
+        var previousContext = pathContext.getContext();
         if (previousContext.isValueNode()) {
             if (previousContext.isNumber()) {
-                throw new IllegalArgumentException(FORMAT_EXCEPTION);
+                throw new IllegalArgumentException(JsonNodePathUtils.FORMAT_EXCEPTION);
             } else {
                 this.applyMask(previousContext.stringValue(), maskField.get(identifyField),
-                        pathPart, (ObjectNode) nodeParent);
+                        pathContext.getPathPart(), (ObjectNode) pathContext.getParent());
             }
         }
-    }
-
-    private JsonNode findArrayParent(JsonNode arrayNodeParent, String pathPart, JsonNode referenceParent) {
-        int index = Integer.parseInt(pathPart);
-        var context = (ArrayNode) arrayNodeParent;
-
-        for (int i = 0; i < context.size(); i++) {
-            if (i == index && !context.get(i).isValueNode()) {
-                referenceParent = context.get(i);
-                return referenceParent;
-            }
-        }
-        return referenceParent;
-    }
-
-    private JsonNode findContext(JsonNode node, String pathPart) {
-        JsonNode context = node.findValue(pathPart);
-        if (context == null) {
-            throw new IllegalArgumentException("\"" + pathPart + "\"" + NOT_FOUND_IN_TREE);
-        }
-        return context;
     }
 
     private void findFieldsForName(JsonNode node, Map<IdentifyField, MaskingFormat> maskField,
@@ -125,18 +77,11 @@ public class JsonSerializer extends StdSerializer<DataMask<?>> {
         }
     }
 
-    private JsonNode convertValue(Object node, JsonMapper mapper) {
-        if (node instanceof String) {
-            return mapper.readTree(node.toString());
-        }
-        return mapper.valueToTree(node);
-    }
-
     private void applyMask(String value, MaskingFormat format, String field, ObjectNode node) {
         try {
             Object resultMask = MaskSerializerCommons.of(format, dataCipher).applyMask(value);
-            if (resultMask instanceof String) {
-                node.put(field, (String) resultMask);
+            if (resultMask instanceof String resultMaskStr) {
+                node.put(field, resultMaskStr);
             } else {
                 node.putPOJO(field, resultMask);
             }
