+++ core/src/main/java/co/com/bancolombia/datamask/databind/unmask/JsonDeserializer.java
@@ -5,6 +5,7 @@ import co.com.bancolombia.datamask.MaskUtils;
 import co.com.bancolombia.datamask.cipher.DataDecipher;
 import co.com.bancolombia.datamask.databind.mask.IdentifyField;
 import co.com.bancolombia.datamask.databind.mask.MaskingFormat;
+import co.com.bancolombia.datamask.databind.util.JsonNodePathUtils;
 import co.com.bancolombia.datamask.databind.util.QueryType;
 import co.com.bancolombia.datamask.databind.util.TransformationType;
 import org.apache.commons.lang3.StringUtils;
@@ -12,7 +13,6 @@ import tools.jackson.core.JacksonException;
 import tools.jackson.core.JsonGenerator;
 import tools.jackson.core.TreeNode;
 import tools.jackson.databind.JsonNode;
-import tools.jackson.databind.ObjectMapper;
 import tools.jackson.databind.SerializationContext;
 import tools.jackson.databind.json.JsonMapper;
 import tools.jackson.databind.node.ArrayNode;
@@ -26,10 +26,7 @@ import java.util.Optional;
 
 public class JsonDeserializer extends StdSerializer<DataUnmasked> {
 
-    public static final String SEPARATOR = "/";
-    public static final String FORMAT_EXCEPTION = "DataMask only support fields in String format";
     public static final String INVALID_CALLER_EXCEPTION = "This method haven't enable for TransformationType.ONLY_MASK";
-    public static final String NOT_FOUND_IN_TREE = " not found in tree.";
 
     private final DataDecipher dataDecipher;
 
@@ -41,7 +38,7 @@ public class JsonDeserializer extends StdSerializer<DataUnmasked> {
     @Override
     public void serialize(DataUnmasked value, JsonGenerator generator, SerializationContext provider) throws JacksonException {
         JsonMapper mapper = new JsonMapper();
-        var objectNode = convertValue(value.getData(), mapper);
+        var objectNode = JsonNodePathUtils.convertValue(value.getData(), mapper);
         Map<IdentifyField, MaskingFormat> values = value.getFields();
         findFields(objectNode, values);
         mapper.writeValue(generator, objectNode);
@@ -49,18 +46,18 @@ public class JsonDeserializer extends StdSerializer<DataUnmasked> {
 
     private void findFields(JsonNode node, Map<IdentifyField, MaskingFormat> maskField) {
 
-        boolean existFieldsWithQueryTypeName = maskField.keySet().stream().anyMatch((identifyField) ->
+        boolean existFieldsWithQueryTypeName = maskField.keySet().stream().anyMatch(identifyField ->
                 identifyField.getQueryType().equals(QueryType.NAME));
         if (existFieldsWithQueryTypeName) {
             this.findFieldsForName(node, maskField, null, node);
         }
-        maskField.keySet().stream().filter((identifyField) ->
-                identifyField.getQueryType().equals(QueryType.PATH)).forEach((identifyField) -> {
+        maskField.keySet().stream().filter(identifyField ->
+                identifyField.getQueryType().equals(QueryType.PATH)).forEach(identifyField -> {
             if (maskField.get(identifyField).getTransformationType().equals(TransformationType.ONLY_MASK)) {
                 throw new IllegalCallerException(INVALID_CALLER_EXCEPTION);
             }
-            String[] querySplit = identifyField.getQuery().split(SEPARATOR);
-            this.findFieldForPath(node, identifyField, maskField, querySplit);
+            String[] querySplit = identifyField.getQuery().split(JsonNodePathUtils.SEPARATOR);
+            this.findFieldForPath(node, querySplit);
         });
     }
 
@@ -96,38 +93,14 @@ public class JsonDeserializer extends StdSerializer<DataUnmasked> {
 
     }
 
-    private void findFieldForPath(JsonNode node, IdentifyField identifyField,
-                                  Map<IdentifyField, MaskingFormat> maskField, String[] querySplit) {
-        var previousContext = node;
-        var nodeParent = node;
-        var pathPart = "";
-        var nodeParentArray = node;
-        var arrayNodeFound = false;
-        var newContext = node;
-        for (String element : querySplit) {
-            pathPart = element;
-            if (!pathPart.isEmpty()) {
-                if (arrayNodeFound) {
-                    nodeParent = findArrayParent(nodeParentArray, pathPart, nodeParent);
-                    previousContext = nodeParent;
-                    newContext = nodeParent;
-                    arrayNodeFound = false;
-                } else {
-                    newContext = findContext(previousContext, pathPart);
-                    previousContext = newContext;
-                }
-                if (!newContext.isValueNode()) {
-                    nodeParent = newContext;
-                }
-                if (newContext.isArray()) {
-                    nodeParentArray = newContext;
-                    arrayNodeFound = true;
-                }
-            }
-        }
+    private void findFieldForPath(JsonNode node, String[] querySplit) {
+        var pathContext = JsonNodePathUtils.navigatePath(node, querySplit);
+        var previousContext = pathContext.getContext();
+        var nodeParent = pathContext.getParent();
+        var pathPart = pathContext.getPathPart();
         if (previousContext.isValueNode()) {
             if (previousContext.isNumber()) {
-                throw new IllegalArgumentException(FORMAT_EXCEPTION);
+                throw new IllegalArgumentException(JsonNodePathUtils.FORMAT_EXCEPTION);
             } else {
                 if (isEncryptedObject(previousContext)) {
                     decipherValueObject(node, pathPart, nodeParent);
@@ -143,27 +116,6 @@ public class JsonDeserializer extends StdSerializer<DataUnmasked> {
         }
     }
 
-    private JsonNode findArrayParent(JsonNode arrayNodeParent, String pathPart, JsonNode referenceParent) {
-        int index = Integer.parseInt(pathPart);
-        var context = (ArrayNode) arrayNodeParent;
-
-        for (int i = 0; i < context.size(); i++) {
-            if (i == index && !context.get(i).isValueNode()) {
-                referenceParent = context.get(i);
-                return referenceParent;
-            }
-        }
-        return referenceParent;
-    }
-
-    private JsonNode findContext(JsonNode node, String pathPart) {
-        JsonNode context = node.findValue(pathPart);
-        if (context == null) {
-            throw new IllegalArgumentException("\"" + pathPart + "\"" + NOT_FOUND_IN_TREE);
-        }
-        return context;
-    }
-
     private void decipherValueObject(JsonNode node, String fieldName, JsonNode parent) {
         ((ObjectNode) parent).put(fieldName,
                 dataDecipher.decipher(node.findValue(DataMaskingConstants.ENCRYPTED_ATTR).asString()));
@@ -173,18 +125,10 @@ public class JsonDeserializer extends StdSerializer<DataUnmasked> {
         ((ObjectNode) parent).put(fieldName, dataDecipher.decipher(value));
     }
 
-    private JsonNode convertValue(Object node, ObjectMapper mapper) {
-        if (node instanceof String) {
-            return mapper.readTree(node.toString());
-        }
-        return mapper.convertValue(node, JsonNode.class);
-    }
-
-
     private boolean isEncryptedString(TreeNode node) {
         return Optional.of(node)
-                .filter(n -> n instanceof StringNode)
-                .map(nodes -> (StringNode) nodes)
+                .filter(StringNode.class::isInstance)
+                .map(StringNode.class::cast)
                 .map(JsonNode::asString)
                 .filter(StringUtils::isNotBlank)
                 .map(t -> t.startsWith(DataMaskingConstants.MASKING_PREFIX))
